Add another layer for historical images
As of this commit the generation of latest images and historical
images is separated. The former are built if `$LATESTRELEASES` is set
to `true` and the latter if `$HISTORICALRELEASES` is set to
`true`. That way we can select what to build by using CI variables.

Conceptually this commit changes the build flow of historical
images. Instead of every flavor downloading the full image (that means
4 downloads of the ISO per release which is extremely load-intensive)
we are caching the extracted ISO in a docker image on its own. That
image is pushed as `TL{YEAR}-historic-iso` and used by all images that
build from the ISO installation.

Unfortunately, GitLab has no easy way to remove this image after the
build, so the ISO images are cached forever, although they are
essentially useless for everyone else.