Incorporate font cache into images
As we discovered in #3, the font cache adds an overhead of approx. 3 MB. It is
not worth adding the complexity of another Dockerfile for this small overhead.
Thus, the cache will always be generated by default. If someone wants to build
the Docker image on their own, there is a build argument to disable cache
generation.