Add a separate stage for the doc-src image
For weeks, the doc-src image failed with no space left on device warnings. This
is not too critical as the image is the least used one but on the other hand, we
needed some kind of bandaid for it as we cannot change GitLab's runners.

Hence, the doc-src image will now be based on the -src image and pull the
documentation from the -doc image, therefore bypassing actual installation
steps. Hopefully, this will resolve the space issues on GitLab.